Move `timestamp_to_local` into a separate utils file
To avoid circular import error from `modal/_output.py`.
